# web/openapi.json
{
  "components": {
    "schemas": {
      "Account": {
        "properties": {
          "cash": {
            "title": "Cash",
            "type": "number"
          },
          "equity": {
            "title": "Equity",
            "type": "number"
          }
        },
        "required": [
          "equity",
          "cash"
        ],
        "title": "Account",
        "type": "object"
      },
      "ApiCall": {
        "properties": {
          "at": {
            "title": "At",
            "type": "string"
          },
          "elapsed_ms": {
            "title": "Elapsed Ms",
            "type": "number"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          }
        },
        "required": [
          "at",
          "method",
          "path",
          "status",
          "elapsed_ms"
        ],
        "title": "ApiCall",
        "type": "object"
      },
...
      "Decision": {
        "properties": {
          "checks": {
            "items": {
              "$ref": "#/components/schemas/RuleCheck"
            },
            "title": "Checks",
            "type": "array"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "regime": {
            "title": "Regime",
            "type": "string"
          },
          "rsi": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Rsi",
            "type": "object"
          },
          "weights": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Weights",
            "type": "object"
          }
        },
        "required": [
          "date",
          "regime",
          "checks",
          "rsi",
          "weights"
        ],
        "title": "Decision",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrderRecord": {
        "properties": {
          "client_order_id": {
            "title": "Client Order Id",
            "type": "string"
          },
          "filled_avg_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filled Avg Price"
          },
          "filled_qty": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filled Qty"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "qty": {
            "title": "Qty",
            "type": "integer"
          },
          "side": {
            "title": "Side",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "side",
          "qty",
          "client_order_id",
          "status"
        ],
        "title": "OrderRecord",
        "type": "object"
      },
...
      "Position": {
        "properties": {
          "market_value": {
            "title": "Market Value",
            "type": "number"
          },
          "qty": {
            "title": "Qty",
            "type": "number"
          }
        },
        "required": [
          "qty",
          "market_value"
        ],
        "title": "Position",
        "type": "object"
      },
      "RuleCheck": {
        "properties": {
          "gap": {
            "title": "Gap",
            "type": "number"
          },
          "left": {
            "title": "Left",
            "type": "string"
          },
          "left_return": {
            "title": "Left Return",
            "type": "number"
          },
          "lookback": {
            "title": "Lookback",
            "type": "integer"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          },
          "right": {
            "title": "Right",
            "type": "string"
          },
          "right_return": {
            "title": "Right Return",
            "type": "number"
          },
          "rule": {
            "title": "Rule",
            "type": "integer"
          }
        },
        "required": [
          "rule",
          "left",
          "right",
          "lookback",
          "left_return",
          "right_return",
          "gap",
          "passed"
        ],
        "title": "RuleCheck",
        "type": "object"
      },
      "Run": {
        "properties": {
          "account": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Account"
              },
              {
                "type": "null"
              }
            ]
          },
          "api_calls": {
            "items": {
              "$ref": "#/components/schemas/ApiCall"
            },
            "title": "Api Calls",
            "type": "array"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Decision"
              },
              {
                "type": "null"
              }
            ]
          },
          "dry_run": {
            "title": "Dry Run",
            "type": "boolean"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "orders": {
            "items": {
              "$ref": "#/components/schemas/OrderRecord"
            },
            "title": "Orders",
            "type": "array"
          },
          "positions": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Position"
            },
            "title": "Positions",
            "type": "object"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/Step"
            },
            "title": "Steps",
            "type": "array"
          },
          "trading_day": {
            "title": "Trading Day",
            "type": "string"
          }
        },
        "required": [
          "trading_day",
          "started_at",
          "finished_at",
          "status",
          "dry_run",
          "error",
          "steps",
          "decision",
          "orders",
          "account",
          "positions",
          "api_calls"
        ],
        "title": "Run",
        "type": "object"
      },
...
      "Step": {
        "properties": {
          "detail": {
            "additionalProperties": true,
            "title": "Detail",
            "type": "object"
          },
          "ms": {
            "title": "Ms",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "status",
          "ms",
          "detail"
        ],
        "title": "Step",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Read-only data behind the web console, published as static JSON files.",
    "title": "algo-trading-bot console API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/decisions/{year}.json": {
      "get": {
        "operationId": "decisions_api_decisions__year__json_get",
        "parameters": [
          {
            "in": "path",
            "name": "year",
            "required": true,
            "schema": {
              "title": "Year",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/Decision"
                  },
                  "title": "Response Decisions Api Decisions  Year  Json Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "strategy() output for each day of a year"
      }
    },
...
    "/api/runs/{day}.json": {
      "get": {
        "operationId": "run_api_runs__day__json_get",
        "parameters": [
          {
            "in": "path",
            "name": "day",
            "required": true,
            "schema": {
              "title": "Day",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One live run with its step trace and orders"
      }
    },
...
  }
}